fix: move fuction to transaction.js
